feat: Refactor code structure and functionality: bucket name and subdir are now driver opts and the volume name can be chosen freely
Motivation: This allows to have multiple volumes pointing to the same bucket and folder. This is important for when updating the volume or its configuration as it allows to create the new volume with the new plugin and connect the services to it before having to remove the old ones.